requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Twitter API
tweepy==4.14.0

# OpenAI API
openai>=1.54.3,<3.0.0

# Environment variables
python-dotenv>=1.1.0

# Image processing
Pillow>=10.1.0

# HTTP requests
requests>=2.32.5,<3.0.0

# Async support
aiohttp>=3.10.11

# Logging
colorlog>=6.8.0
